renamed files to be more pythonic
